Move file parsing responsibility to loader the instance
