+++ src/MageTest/Manager/Attributes/Provider/Loader/ParseFields.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace MageTest\Manager\Attributes\Provider\Loader;
+
+
+/**
+ * Interface ParseFields
+ *
+ * @package MageTest\Manager\Attributes\Provider\Loader
+ */
+interface ParseFields
+{
+    /**
+     * @param $model
+     * @return mixed
+     */
+    public function parseFields($model);
+}
