+++ src/MageTest/Manager/Attributes/Provider/Loader/YmlLoader.php
@@ -5,18 +5,68 @@ namespace MageTest\Manager\Attributes\Provider\Loader;
 
 use Symfony\Component\Yaml\Yaml;
 
-class YmlLoader implements Loader
+/**
+ * Class YmlLoader
+ *
+ * @package MageTest\Manager\Attributes\Provider\Loader
+ */
+class YmlLoader implements Loader, ParseFields
 {
 
+    /**
+     * @var \Symfony\Component\Yaml\Yaml
+     */
     private $yaml;
 
+    /**
+     * @param Yaml $yaml
+     */
     public function __construct(Yaml $yaml = null)
     {
         $this->yaml = $yaml ? : new Yaml;
     }
 
+    /**
+     * @param $file
+     * @return array
+     */
     public function load($file)
     {
         return $this->yaml->parse($file);
     }
+
+    /**
+     * @param $model
+     * @return array
+     */
+    public function parseFields($model)
+    {
+        return array(
+            $this->parseModel(key($model)) => array(
+                'depends' => $this->parseDependencies(key($model)),
+                'attributes' => reset($model)
+            )
+        );
+    }
+
+    /**
+     * @param $key
+     * @return mixed
+     */
+    private function parseModel($key)
+    {
+        preg_match("/[^ (]+/", $key, $matches);
+        return $matches[0];
+    }
+
+    /**
+     * @param $key
+     * @return mixed
+     */
+    private function parseDependencies($key)
+    {
+        preg_match("/ \((.*)\)/", $key, $matches);
+        return $matches[1];
+    }
+
 }
